fix 'bad checksum' issue.
When a second client used to connect to a room, it was receiving the
last state of the room, instead of the last snapshot state other users
already have.

After receiving the first patch broadcast, the new client used to have a
'bad checksum' error in the client-side, because the patch used the
previous sent state snapshot.

TLDR: The clients must share a common snapshot to be able to apply the
patch.